Cross-platform MsQuic+wolfSSL: native crypto, Linux Docker build, UI improvements
- Add crypt_wolfssl.c: native wolfcrypt crypto backend replacing both
  crypt_openssl.c (Linux) and crypt_bcrypt.c (Windows) — AES-GCM,
  ChaCha20-Poly1305, HMAC, SP 800-108 KBKDF via wolfcrypt API directly
- Strip wolfssl_compat.h to TLS-only stubs (removed all crypto shims)
- Fix MsQuic static archive merge on Linux (exclude system libs m/pthread/rt)
- Add docker-compose-build.yml and Linux cmake preset for server builds
- Fix platform portability: Sleep() → std::this_thread::sleep_for,
  guard Windows crash handler behind #ifdef _WIN32
- UI: context menu, lobby model improvements, sound player updates
- Protocol: add screen share message type

+++ common/src/quic_common.cpp
@@ -10,7 +10,7 @@ const QUIC_API_TABLE* quic_init() {
 
     QUIC_STATUS status = MsQuicOpen2(&g_quic_api);
     if (QUIC_FAILED(status)) {
-        std::fprintf(stderr, "[QUIC] MsQuicOpen2 failed: 0x%lx\n", status);
+        std::fprintf(stderr, "[QUIC] MsQuicOpen2 failed: 0x%lx\n", (unsigned long)status);
         return nullptr;
     }
